Vary N (# total commercial trips)

Function to run Cochran estimator with a variable big N (total number of trips). Sequences from the numnber of observed trips up to a maximum number of total trips. N > n

vary.bigN(bdat, bigN, by = 10)

Arguments

bdat
data loaded from get.bdat

bigN
Total trips

by
sequencing step. Start is the number of observed trips, end is big N

Value

SD and CV from cochran_rse

Examples

vary.bigN

#> function (bdat, bigN, by = 10) 
#> {
#>     N = data.frame(N = seq(nrow(bdat), bigN, by = by))
#>     s = ddply(N, 1, function(x) cochran_rse(bdat, x))
#>     names(s)[4:5] = c("RE_se", "RE_rse")
#>     s
#> }
#> <bytecode: 0x0000000030f34f18>
#> <environment: namespace:discaRd>